feat: add schema infer types
