+++ examples/postgres/src/db/schema.ts
@@ -1,10 +1,13 @@
 import { model } from "kineo";
+import type { InferModel } from "kineo/schema";
 
 export const user = model((s) => ({
   id: s.int().id(),
 })).relate((s) => ({
   posts: s.relation(post).many().required(),
 }));
+
+export type User = InferModel<typeof user>;
 
 export const post = model((s) => ({
   id: s.int().id(),
@@ -19,3 +22,5 @@ export const post = model((s) => ({
       .required(),
   }))
   .index(["authorId"]);
+
+export type Post = InferModel<typeof post>;
